// src/lib/types.ts
export type RefinedCast = {
  id: string;
  name: string;
  poster: string | null;
  character: string;
};

export type RefinedCrew = {
  name: string;
  id: string;
};

export type RefinedCollection = {
  id: string;
  name: string;
  poster: string;
  backdrop: string;
};

export type RefinedMovieData = {
  awards: string;
  backdrop: string;
  cast: RefinedCast[];
  collection: RefinedCollection | null;
  country: string[];
  directors: RefinedCrew[];
  genres: string[];
  homepage: string;
  imdb_id: string;
  imdb_rating: string;
  language: string;
  languages: string[];
  original_title: string;
  plot: string;
  poster: string;
  production_companies: {
    name: string;
    id: string;
  }[];
  rated: string;
  release_date: number;
  runtime: string;
  status: string;
  overview: string;
  tagline: string;
  title: string;
  tmdb_id: string;
  tmdb_rating: string;
  trailers: { name: string; key: string }[];
  writers: RefinedCrew[];
  year: number;
};